@@ -16,8 +16,8 @@ public sealed class PaddedButton : UIButton
 		TopPadding = topPadding;
 		RightPadding = rightPadding;
 		BottomPadding = bottomPadding;
-
-		ContentEdgeInsets = new UIEdgeInsets(topPadding, leftPadding, bottomPadding, rightPadding);
+		
+		SetPadding(leftPadding, topPadding, rightPadding, bottomPadding);
 	}
 
 	/// <summary>
@@ -39,4 +39,19 @@ public sealed class PaddedButton : UIButton
 	/// Bottom Padding
 	/// </summary>
 	public NFloat BottomPadding { get; }
+	
+	void SetPadding(NFloat leftPadding, NFloat topPadding, NFloat rightPadding, NFloat bottomPadding)
+	{
+		if (OperatingSystem.IsIOSVersionAtLeast(15))
+		{
+			if (Configuration is not null)
+			{
+				Configuration.ContentInsets = new NSDirectionalEdgeInsets(topPadding, leftPadding, bottomPadding, rightPadding);
+			}
+		}
+		else
+		{
+			ContentEdgeInsets = new UIEdgeInsets(topPadding, leftPadding, bottomPadding, rightPadding);
+		}
+	}
 }
